gettledogbot/pyproject.toml
[tool]
[tool.poetry]
name = "root"
version = "0.0.0"

[tool.poetry.dependencies]
python = "*"
"discord.py" = "*"
flask = "*"
